renamed some of the namespaces and correct some namepsace information that was missing
