+++ OPC UA Nodeset WebAPI/Controllers/v1/PropertyController.cs
@@ -6,7 +6,7 @@ using OPC_UA_Nodeset_WebAPI.UA_Nodeset_Utilities;
 using System;
 using System.Web;
 
-namespace OPC_UA_Nodeset_WebAPI.api.v1.Controllers
+namespace OPC_UA_Nodeset_WebAPI.Controllers.v1
 {
     [ApiController]
     [Route("api/v1/property")]
