+++ OPC UA Nodeset WebAPI/Model/v1/Responses/DataTypeResponse.cs
@@ -2,6 +2,7 @@
 using Opc.Ua.Export;
 using OPC_UA_Nodeset_WebAPI.UA_Nodeset_Utilities;
 using System.Xml.Linq;
+using Opc.Ua.Export.v1.Responses;
 
 namespace OPC_UA_Nodeset_WebAPI.Model.v1.Responses
 {
